Fix error, add option to specify output file path
It seems with the newer version of PrusaSlicer, they don't include the
number on that last "toolchange". It doesn't actually do a toolchange so
maybe that's why. Either way, that was being used to determine if it was
the correct place to remove that last ram section of gcode. This change
goes with a different approach (turns out to be a much simpler approach)
for calculating the count of a particular toolchange.

This also adds the option to specify the output file path. This is done
in some of the newer PostProcessing scripts and its convenient when
running the script manually (in particular, while doing development).