Fix seam backup
It was computing the extrusion incorrectly. This was because I didn't
realize that there can be a non-travel/non-printing G1 at the end of a
line that just updates the speed. We now ignore that line when
determining the back up path.

Also updated the 'wipe_point' function name to 'back_up_point' to be
more accurate as to what it is doing (Also changed the math around in
the function, gets to the same numbers but just in a way that I think is
a little more clear for a "back up").

Also keep track of a line's wipe start and end (the usual wipe, the one
created by the slicer).