Add option to only close perimeter extrusions
Add a flag to only close perimeters rather than all extrusions. This is
now the default. It does seem to make the most sense for a typical
print for this to be the default. If the old behavior of closing gaps
on all types of extrusions is desired, pass the '--no-perimeters-only'
flag.